#82 refactor : ItemService와 ItemQueryService를 하나의 ItemService로 관리하게끔 변경
클래스단에 붙어있는 @Transactional(readOnly = true)로 메서드에@Transactional(readOnly = true) 붙여져있는 어노테이션 삭제
